+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * `LiteralInInterpolation` cop does auto-correction. ([@tmr08c][])
 * [#1865](https://github.com/bbatsov/rubocop/issues/1865): New cop `Lint/UnneededDisable` checks for `rubocop:disable` comments that can be removed. ([@jonas054][])
 * `EmptyElse` cop does auto-correction. ([@lumeet][])
+* Show reference links when displaying style guide links. ([@rrosenblum][])
 
 ### Changes